refactor(slurm): split runtimes/slurm.py 639 LOC → focused 6-file package
No behavior change. Carve runtimes/slurm.py into a package:

  runtimes/slurm/
    __init__.py     67 LOC  re-exports + test-monkeypatch surface
    _constants.py   36 LOC  hardener strings (REQUIRED_SHEBANG, …)
    _state.py       56 LOC  per-agent JSON state file IO
    _heartbeat.py   81 LOC  compute-node heartbeat shell-fragment
    _render.py     241 LOC  sbatch directive + script template rendering
    _runtime.py    286 LOC  SlurmRuntime class + scitex-hpc dual-write

Backward-compat: all current imports keep working —
  from scitex_agent_container.runtimes.slurm import SlurmRuntime
  from scitex_agent_container.runtimes.slurm import REQUIRED_SHEBANG
  ...

Tests: 52/53 pass (same 1 pre-existing failure as develop baseline,
unrelated to this refactor — `local_state.runtime_path()` caches the
base path at import time, so `monkeypatch.setenv('HOME', tmp_path)`
in test_start_writes_sbatch_script_file doesn't take effect).

Original slurm.py renamed to _runtime.py via `git mv` first to
preserve blame on the bulk of the code; helpers carved out from
there into the sibling files.
