+++ src/scitex_agent_container/runtimes/slurm/_render.py
@@ -1,0 +1,241 @@
+"""Sbatch wrapper script rendering.
+
+Builds the wrapper-script body from an :class:`AgentConfig`. The returned
+text is suitable for writing to a file and submitting via
+``sbatch <file>``. All hardeners and plugin hook ports are applied.
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+from ...config import AgentConfig
+from ._constants import (
+    REQUIRED_EXIT_TRAP_MARKER,
+    REQUIRED_SHEBANG,
+    REQUIRED_STRICT_MODE,
+    REQUIRED_USR1_TRAP_MARKER,
+    REQUIRED_XTRACE,
+)
+from ._heartbeat import _heartbeat_block
+from ._state import _read_state
+
+
+def _build_claude_command(cfg: AgentConfig) -> str:
+    """Build the claude-code invocation from the agent's claude spec.
+
+    Mirrors the shape ClaudeCodeRuntime uses for its tmux session. Kept
+    deliberately minimal and consumer-agnostic: any consumer-specific flags
+    (e.g. ``--dangerously-load-development-channels``) are passed in via
+    ``config.claude.flags`` from the agent YAML. ``claude`` is resolved
+    via ``PATH`` — HPC hooks can prepend ``module load nodejs`` etc.
+    """
+    parts = ["claude"]
+
+    session_mode = getattr(cfg.claude, "session", "continue-or-new")
+    if session_mode in ("continue", "continue-or-new"):
+        parts.append("--continue")
+
+    model = cfg.model or "sonnet"
+    parts.extend(["--model", model])
+
+    flags = list(getattr(cfg.claude, "flags", []) or [])
+    for flag in flags:
+        parts.append(str(flag))
+
+    # Command is embedded inside a double-quoted tmux new-session argument.
+    return " ".join(parts)
+
+
+def _hook_source(path: str, phase: str, agent_id: str, logs_dir: str) -> str:
+    """Emit a shell fragment that sources ``path`` if it exists.
+
+    Empty ``path`` emits an empty string so the wrapper stays clean.
+    """
+    if not path:
+        return ""
+    return (
+        f"\n# Hook: {phase}\n"
+        f'if [[ -f "{path}" ]]; then\n'
+        f'    SAC_AGENT_ID="{agent_id}" \\\n'
+        f'    SAC_WORKDIR="${{SAC_WORKDIR:-}}" \\\n'
+        f'    SAC_LOG_FILE="{logs_dir}/${{SLURM_JOB_ID:-nojob}}.out" \\\n'
+        f'    SAC_JOB_ID="${{SLURM_JOB_ID:-}}" \\\n'
+        f'    SAC_PHASE="{phase}" \\\n'
+        f'    source "{path}"\n'
+        f"fi\n"
+    )
+
+
+def _sbatch_directives(cfg: AgentConfig) -> list[str]:
+    slurm = cfg.slurm
+    job_name = slurm.job_name or cfg.name
+    # Expand ~ here: bash does NOT expand ~ inside double-quoted strings,
+    # and every path in the emitted script is double-quoted for safety.
+    logs_dir = str(Path(slurm.logs_dir).expanduser())
+
+    directives = [f"#SBATCH --job-name={job_name}"]
+    if slurm.partition:
+        directives.append(f"#SBATCH --partition={slurm.partition}")
+    directives.extend(
+        [
+            f"#SBATCH --time={slurm.time_limit}",
+            f"#SBATCH --nodes={slurm.nodes}",
+            f"#SBATCH --ntasks={slurm.ntasks}",
+            f"#SBATCH --cpus-per-task={slurm.cpus_per_task}",
+            f"#SBATCH --mem={slurm.mem}",
+        ]
+    )
+    if slurm.gres:
+        directives.append(f"#SBATCH --gres={slurm.gres}")
+    if slurm.signal:
+        directives.append(f"#SBATCH --signal={slurm.signal}")
+    directives.append(f"#SBATCH --output={logs_dir}/%x_%j.out")
+    directives.append(f"#SBATCH --error={logs_dir}/%x_%j.err")
+    directives.extend(slurm.extra_directives)
+    return directives
+
+
+def render_sbatch_script(cfg: AgentConfig) -> str:
+    """Render the full sbatch wrapper script as a string.
+
+    The returned text is suitable for writing to a file and submitting via
+    ``sbatch <file>``. All hardeners and plugin hook ports are applied.
+    """
+    slurm = cfg.slurm
+    agent_id = cfg.name
+    # Expand ~ at render time — bash does not expand ~ inside the
+    # double-quoted strings we emit into the wrapper (cd "$workdir", exec
+    # > "$logs_dir/...", hook SAC_* env).
+    workdir = str(
+        Path(
+            cfg.workdir or f"~/.scitex/agent-container/workspaces/{agent_id}"
+        ).expanduser()
+    )
+    logs_dir = str(Path(slurm.logs_dir).expanduser())
+    tmux_session = cfg.screen_name or agent_id
+    claude_cmd = _build_claude_command(cfg)
+
+    directives = "\n".join(_sbatch_directives(cfg))
+
+    pre_submit = _hook_source(slurm.hooks.pre_submit, "pre_submit", agent_id, logs_dir)
+    pre_agent = _hook_source(slurm.hooks.pre_agent, "pre_agent", agent_id, logs_dir)
+    walltime_hook = _hook_source(
+        slurm.hooks.walltime_signal, "walltime_signal", agent_id, logs_dir
+    )
+    post_agent = _hook_source(slurm.hooks.post_agent, "post_agent", agent_id, logs_dir)
+    heartbeat_block = _heartbeat_block(cfg, logs_dir)
+
+    resubmit_line = (
+        'sbatch "$0"'
+        if slurm.auto_resubmit
+        else 'echo "[sac/slurm] auto_resubmit disabled"'
+    )
+
+    script = f"""{REQUIRED_SHEBANG}
+{directives}
+#
+# Generic sac SLURM wrapper. Generated by
+# scitex_agent_container.runtimes.slurm.render_sbatch_script — do not
+# hand-edit on the compute host; regenerate via ``sac render-sbatch``.
+
+{REQUIRED_STRICT_MODE}
+
+# pre_submit runs BEFORE the wrapper continues into the job body. When this
+# script is invoked under sbatch, SLURM_JOB_ID is already set by SLURM — so
+# in practice pre_submit and pre_agent run on the same node (the compute
+# node). External orchestrators who need a true "login-node pre-submit"
+# should perform that side-effect in their own submit wrapper before
+# calling sbatch. This hook still runs here for symmetry with the phase
+# taxonomy.
+{pre_submit}
+mkdir -p "{logs_dir}"
+exec > "{logs_dir}/${{SLURM_JOB_ID:-nojob}}.out" 2>&1
+{REQUIRED_XTRACE}
+
+# Fail loud on drop-through: if we ever exit the hold below unexpectedly,
+# this trap surfaces the return code rather than the scheduler silently
+# reaping the job. Also tears down the optional heartbeat daemon so
+# compute-node pushers never outlive the allocation.
+trap 'rc=$?; kill "${{SAC_HEARTBEAT_PID:-0}}" 2>/dev/null || true; echo "{REQUIRED_EXIT_TRAP_MARKER} rc=$rc at $(date -u +%FT%TZ)" >&2; exit "${{rc:-1}}"' EXIT
+
+# Walltime auto-resubmit: SLURM fires SIGUSR1 ``@3600`` seconds before
+# walltime (see --signal directive). The handler sources the external
+# walltime_signal hook (for e.g. hub notifications) and then resubmits
+# this exact script so the allocation is seamlessly replaced.
+{REQUIRED_USR1_TRAP_MARKER}() {{
+    echo "[sac/slurm] walltime signal received at $(date -u +%FT%TZ)"
+    {walltime_hook.strip() if walltime_hook.strip() else "true  # no walltime_signal hook"}
+    {resubmit_line}
+}}
+trap {REQUIRED_USR1_TRAP_MARKER} USR1
+
+export SAC_AGENT_ID="{agent_id}"
+export SAC_WORKDIR="{workdir}"
+mkdir -p "{workdir}"
+cd "{workdir}"
+
+# pre_agent hook mutates env (module load, LD_LIBRARY_PATH, consumer-
+# specific exports like SCITEX_OROCHI_*). Sourced so the exports survive
+# into the agent process.
+{pre_agent}
+# tmux socket lives under HOME to survive shared /tmp cleanups.
+export TMUX_TMPDIR="${{HOME}}/.tmux-sockets"
+mkdir -p "${{TMUX_TMPDIR}}"
+
+# Clean up any stale tmux server from a previous allocation on this node.
+tmux kill-server 2>/dev/null || true
+sleep 1
+
+# Spawn the agent in a detached tmux session. This returns immediately;
+# the hold at the bottom of the script is what actually keeps the SLURM
+# allocation alive.
+tmux new-session -d -s "{tmux_session}" "{claude_cmd}"
+{heartbeat_block}
+# post_agent hook fires once the tmux session ends. The hold below keeps
+# the job alive even if the agent dies — external observers (healers)
+# can inspect and decide whether to scancel.
+{post_agent}
+
+# Persistent hold. Idempotent, never decrements — safer than a literal
+# sleep computed off walltime (which can race-terminate near the
+# boundary).
+echo "[sac/slurm] entering persistent hold at $(date -u +%FT%TZ)"
+{slurm.hold}
+"""
+    return script
+
+
+def render_attach_command(cfg: AgentConfig, job_id: str | None = None) -> str:
+    """Return a shell command that attaches to the running agent's tmux.
+
+    Uses ``srun --jobid=$JID --pty bash -c 'tmux attach -t <session>'``
+    so the operator lands on the compute node running the agent.
+
+    If ``job_id`` is not given, reads from sac's slurm state file.
+    """
+    if job_id is None:
+        state = _read_state(cfg.name) or {}
+        job_id = state.get("job_id", "")
+    session = cfg.screen_name or cfg.name
+
+    attach_hook = cfg.slurm.hooks.attach
+    pre = ""
+    if attach_hook:
+        pre = (
+            f'SAC_AGENT_ID="{cfg.name}" SAC_JOB_ID="{job_id}" '
+            f'SAC_PHASE="attach" bash "{attach_hook}" && '
+        )
+    return (
+        f"{pre}srun --jobid={job_id} --pty bash -c "
+        f"'tmux -L default attach -t {session}'"
+    )
+
+
+__all__ = [
+    "_build_claude_command",
+    "_hook_source",
+    "_sbatch_directives",
+    "render_attach_command",
+    "render_sbatch_script",
+]
